Move InstallButton to new file
